feat: 토큰 없는 route hanlder 함수 분리
